+++ __init__.py
@@ -4,7 +4,7 @@ bl_info = {
     "name": "Storytools - Storyboard Tools",
     "description": "Set of tools for Storyboarding",
     "author": "Samuel Bernou",
-    "version": (1, 1, 1),
+    "version": (1, 2, 0),
     "blender": (3, 3, 0),
     "location": "View3D",
     "warning": "",
@@ -25,8 +25,8 @@ from . import GZ_toolbar
 from . import OP_workspace_setup
 from . import handles
 from . import panels
+from . import keymaps
 from .fn import get_addon_prefs
-# from . import keymaps
 
 modules = (
     OP_git_update,
@@ -40,6 +40,7 @@ modules = (
     GZ_toolbar,
     OP_workspace_setup,
     panels,
+    keymaps,
 )
 
 def register():